% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_occurrance.R
\name{amino_acid_occurance}
\alias{amino_acid_occurance}
\title{Histogram for amino acid occurance}
\usage{
amino_acid_occurance(amino_acid_seq)
}
\arguments{
\item{amino_acid_seq}{}
}
\description{
Histogram for amino acid occurance
}
