% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_unique_chars.R
\name{barplot_unique_chars}
\alias{barplot_unique_chars}
\title{Barplot for unique characters}
\usage{
barplot_unique_chars(aa_sequence)
}
\arguments{
\item{aa_sequence}{this is a string of characters, in our case a sequence of amino acids}
}
\value{
The function will return a barplot showing occurence of each character or amino acid
}
\description{
Barplot for unique characters
}
