% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonextract.R
\name{codon_extract}
\alias{codon_extract}
\title{Codon Extract}
\usage{
codon_extract(sequenceDNA, start = 1)
}
\value{
Input string separated into codons
}
\description{
Codon Extract
}
\examples{
codon_extract("ATGAAAATA")
# output: "ATG AAA ATA"
}
