% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonextract.R
\name{codon_extractor}
\alias{codon_extractor}
\title{Codon Extract}
\usage{
codon_extractor(concatenated_string, start = 1)
}
\arguments{
\item{concatenated_string}{string containing DNA sequence}
}
\value{
Input string separated into codons
}
\description{
Codon Extract
}
\examples{
codon_extract("ATGAAAATA")
# output: "ATG AAA ATA"
}
