% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conversion_table}
\alias{conversion_table}
\title{RNA to amino acid conversion table}
\format{
A list of 64 characters
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
conversion_table
}
\description{
This is a table which contains the conversion values of RNA triplets to the corresponding amino acids
}
\keyword{datasets}
