% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_codons.R
\name{creating_codons}
\alias{creating_codons}
\title{Codons creation from RNA sequence}
\usage{
creating_codons(RNA_seq, start = 1)
}
\arguments{
\item{start}{}
}
\value{
codons
}
\description{
Codons creation from RNA sequence
}
