% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_seq.R
\name{generate_random_seq}
\alias{generate_random_seq}
\title{Generating random DNA sequence}
\usage{
generate_random_seq(DNA_seq)
}
\arguments{
\item{DNA_seq}{}
}
\value{

}
\description{
Generating random DNA sequence
}
