% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occurrence.R
\name{plot_occurrence}
\alias{plot_occurrence}
\title{Plot the occurrences within an amino acid sequence}
\usage{
plot_occurrence(amino_acid_sequence)
}
\arguments{
\item{amino_acid_sequence}{A character string containing a sequence of amino acids.}
}
\value{
A histogram plot of the occurence of each unique amino acid.
}
\description{
Plot the occurrences within an amino acid sequence
}
\examples{
sequence <- "ACDEFGHIKLMNPQRSTVWYACDEFGHIKLMN"
plot_occurrence(sequence)

}
