% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_DNA.R
\name{random_DNA}
\alias{random_DNA}
\title{Random DNA String Creator of Input Length}
\usage{
random_DNA(string_length)
}
\arguments{
\item{string_length}{input length of the random DNA}
}
\value{
the concatenated sequence string of DNA
}
\description{
Random DNA String Creator of Input Length
}
\examples{
random_DNA(3)
# possible output: "ACT"

random_DNA(5)
# possible output: "TCAGT"
}
