% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\docType{data}
\name{random_sequence}
\alias{random_sequence}
\title{Title Transcription Function}
\format{
An object of class \code{character} of length 1.
}
\usage{
random_sequence
}
\arguments{
\item{concatenated_string}{A string of DNA}
}
\value{
Returns the transcribed string of DNA in the form of an RNA string
}
\description{
Title Transcription Function
}
\keyword{datasets}
