% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_nucleotides.R
\name{sample_nucleotides}
\alias{sample_nucleotides}
\title{Create A Nucleotide Sample}
\usage{
sample_nucleotides(sample_size)
}
\arguments{
\item{sample_size}{size of desired sample (an integer)}
}
\value{
a sample of nucleotides of size sample_size
}
\description{
Create A Nucleotide Sample
}
\examples{
sample <- sample_nucleotides(10)
}
