% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{transcription}
\alias{transcription}
\title{Title: DNA Transcription Function}
\usage{
transcription(concatenated_string)
}
\arguments{
\item{concatenated_string}{A string of DNA created in function 1}
}
\value{
A string of RNA that has been transcribed from the input DNA sequence
}
\description{
Title: DNA Transcription Function
}
