% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translate_rna}
\alias{translate_rna}
\title{Translate Codons to Amino Acids}
\usage{
translate_rna(codons)
}
\arguments{
\item{codons}{A character vector where each element is a codon (a triplet of RNA bases).}
}
\value{
A character string of the amino acid sequence.
}
\description{
This function translates a series of RNA codons into an amino acid chain using a codon table.
}
\examples{
translate_rna(c("AUG", "GCC", "AUG", "GCG"))
}
