% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{translation from RNA sequence codons to amino acid sequence}
\usage{
translation(codons)
}
\arguments{
\item{codons}{}
}
\value{
codons_concat
}
\description{
translation from RNA sequence codons to amino acid sequence
}
