% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_extractor.R
\name{codon_extractor}
\alias{codon_extractor}
\title{Codon_extractor}
\usage{
codon_extractor(input_sequence, start = 1)
}
\arguments{
\item{input_sequence}{a string containing a DNA/RNA sequence}
}
\value{
a string of codons constructed from input sequence
}
\description{
Codon_extractor
}
\examples{
# codon_extractor("ATTGATTTA")
# output: "ATT GAT TTA"

}
