% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occurrence.R
\name{plot_occurrence}
\alias{plot_occurrence}
\title{Plot the occurrences within an amino acid sequence}
\usage{
plot_occurrence(amino_acid_sequence)
}
\arguments{
\item{amino_acid_sequence}{A character string containing a sequence of amino acids.}
}
\value{
A barplot of the occurence of each unique amino acid.
}
\description{
Plot the occurrences within an amino acid sequence
}
\examples{
sequence <- paste0(sample(c("A", "C", "D", "E", "F", "G", "H",
                                      "I", "K", "L", "M", "N", "P", "Q",
                                      "R", "S", "T", "V", "W", "Y"),
                                      1000, replace = TRUE), collapse = "")

plot_occurrence(sequence)

}
