% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_to_codons.R
\name{seq_to_codons}
\alias{seq_to_codons}
\title{seq_to_codons}
\usage{
seq_to_codons(nucleotide_seq, start = 1)
}
\arguments{
\item{nucleotide_seq}{String/character of nucleotide sequence}

\item{start}{Position of the first nucleotide to part into codons. Default is pos 1, parting the whole sequence into codons.}
}
\value{
Takes a string/character of nucleotide sequence and returns a list of codons in the sequence. You can chose to start the slicing of the sequence after the 1st position, if defining new value for start.
}
\description{
seq_to_codons
}
\examples{
nucleotide_seq <- "gaccagatcgataggcggcaatttgaacgctgagaacatcatccggatgtccgctcccgtgtttgtctcagctttgtgcagtcagtgggagggtgttcccaacgactctccgccttttgacttctaaagggttctaggcttccaatcatgtcgccccagcttttccgagaacgatggtattttcggggggggcttcacgttctgtggagaatatcactgtatgtcgatctaggacggcaacttctggaatcaccgattacgtcggcagactgtcatacctgaaacggattccagacgcatagacggcttgcatctggggcggcagaaatatggtcgagactccgacccgagcttgcgggc"
seq_to_codons(nucleotide_seq)
}
