% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{Title Transcription Function}
\usage{
transcription(concatenated_string)
}
\arguments{
\item{concatenated_string}{A string of DNA}
}
\value{
Returns the transcribed string of DNA in the form of an RNA string
}
\description{
Title Transcription Function
}
\examples{
random_sequence <- paste(sample(c("A", "C", "G", "U"),
                                          50, replace = TRUE),
                                          collapse = "")
transcription(random_sequence)
}
